Updated definition handling
new: Add lifecycle callback methods for Definition. This should enable advanced custom definitions
refactor: Simplify the definitions list transfer between the Model and the Statement. There is no custom method for it, the default Statement  get it from the constructor as a list
new: Add transaction support for the default Statement. It can revert the internal state if something goeas wrong
update: To simplify the definition handling, each base definition behavior is written in the Statement instead, the Definition in this case is just a data holder. Along with that default behavior the system supports custom definitions like before (in fact, a more
advanced way)
new: Implement `Definition\FieldForeign` create, update and delete operations
remove: `Model\ItemInterface` and `Model\Item` removed and their usage replaced with associative arrays. It's better for performance, and View-s (and the Model) can be used for the same functionality
refactor: Rename and relocate methods to show the functionality better
update: Documentation changes, fixes